# dashboard/docs/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "IT Global Dashboard API",
    "description": "IT Global 비즈니스 관리 대시보드를 위한 종합 API입니다.\n\n## 주요 기능\n- 프로젝트 관리 및 추적\n- 매출 및 재무 분석\n- 실시간 모니터링 및 메트릭\n- 사용자 관리 및 인증\n- 캐시 및 시스템 관리\n\n## 인증 방식\n이 API는 여러 인증 방법을 지원합니다:\n- 세션 기반 인증 (웹 인터페이스)\n- Bearer 토큰 인증 (API 접근)\n- Google OAuth 2.0 (웹 인터페이스)\n\n## 표준 응답 형식\n모든 API 응답은 다음과 같은 표준 형식을 따릅니다:\n```json\n{\n  \"success\": true|false,\n  \"data\": any,\n  \"error\": {...},\n  \"meta\": {...}\n}\n```\n\n## 버전 관리\nAPI는 URL 경로 기반 버전 관리를 사용합니다 (/api/v1/, /api/v2/ 등).",
    "version": "1.0.0",
    "contact": {
      "name": "IT Global Development Team",
      "email": "dev@itglobal.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://dashboard.itglobal.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/versions": {
      "get": {
        "operationId": "list_api_versions",
        "summary": "모든 API 버전 목록 조회",
        "description": "summary: List all API versions\ndescription: Get information about all available API versions\nresponses:\n200:\ndescription: List of API versions",
        "tags": [
          "API Versioning"
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/version/current": {
      "get": {
        "operationId": "current_api_version",
        "summary": "현재 요청의 API 버전 정보",
        "description": "summary: Get current API version\ndescription: Get information about the API version being used for this request\nresponses:\n200:\ndescription: Current API version information",
        "tags": [
          "API Versioning"
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardResponse"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "인증 필요",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "StandardResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청 성공 여부"
          },
          "data": {
            "description": "응답 데이터"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetails"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "success",
          "data",
          "error",
          "meta"
        ]
      },
      "ErrorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/StandardResponse"
          },
          {
            "type": "object",
            "properties": {
              "success": {
                "type": "boolean",
                "example": false
              },
              "data": {
                "type": "null"
              }
            }
          }
        ]
      },
      "ErrorDetails": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "에러 코드"
          },
          "message": {
            "type": "string",
            "description": "에러 메시지"
          },
          "details": {
            "type": "object",
            "description": "추가 에러 정보"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ResponseMeta": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "응답 시간"
          },
          "version": {
            "type": "string",
            "description": "API 버전"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "description": "요청 ID"
          }
        },
        "required": [
          "timestamp",
          "version",
          "request_id"
        ]
      }
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  },
...
}